requirements.txt
streamlit
python-dotenv
groq
textblob
pandas
matplotlib
